pyproject.toml
[build-system]
requires = [
    "hatchling>=1,<2",
    "hatch-vcs>=0.5",
]
build-backend = "hatchling.build"

[project]
name = "tap-navan"
dynamic = ["version"]
description = "Singer tap for Navan, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Edgar Ramírez-Mondragón", email = "emondragon@meltano.com" }]
keywords = [
    "ELT",
    "Navan",
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
license = "Apache-2.0"
license-files = [ "LICENSE" ]
requires-python = ">=3.10"
dependencies = [
    "singer-sdk~=0.53.6",
    "requests~=2.33.0",
    "typing-extensions>=4.5.0; python_version < '3.13'",
]

[project.optional-dependencies]
s3 = [
    "s3fs>=2026.1.0",
]

[project.scripts]
# CLI declaration
tap-navan = 'tap_navan.tap:TapNavan.cli'

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "pytest>=9",
    "pytest-github-actions-annotate-failures>=0.4",
    "python-dotenv>=1.2.2",
    "singer-sdk[testing]",
]
typing = [
    "mypy>=1.19.1",
    "ty>=0.0.18",
    "types-requests",
]

[tool.hatch.version]
fallback-version = "0.1.0.dev0"
source = "vcs"

[tool.mypy]
enable_error_code = [
    "ignore-without-code",
    "redundant-expr",
    "truthy-bool",
]
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.ty.rules]
unused-type-ignore-comment = "ignore"

[tool.pytest]
addopts = [
    "-ra",
]
filterwarnings = [
    "error",
    "once:No records were available to test:UserWarning",
    "once:Fields in transformed catalog but not in records:UserWarning",
]
log_level = "INFO"
minversion = "9"
strict = true
testpaths = [
    "tests",
]

[tool.ruff]
line-length = 100
required-version = ">=0.15"

[tool.ruff.lint]
future-annotations = true
ignore = [
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

# This configuration can be used to customize tox environments
[tool.tox]
requires = [
    "tox>=4.43",
    "tox-uv",
]
env_list = [
    "typing",
    { product = [
        { prefix = "3.", start = 10 },
    ] },
]

[tool.tox.env_run_base]
description = "run tests on Python {env_name}"
runner = "uv-venv-lock-runner"
pass_env = [
    "GITHUB_*",
    "TAP_NAVAN_*",
]
dependency_groups = [ "test" ]
commands = [ [ "pytest", { replace = "posargs", default = [ "tests" ], extend = true } ] ]

[tool.tox.env.typing]
description = "type checks"
dependency_groups = [ "test", "typing" ]
commands = [
    [ "mypy", { replace = "posargs", default = [ "tap_navan", "tests" ], extend = true } ],
    [ "ty", "check", { replace = "posargs", default = [ "tap_navan", "tests" ], extend = true } ],
]
